Create solved.jsx
the ErrorMessage component contains a ScrollView component that wraps a View component with a defined width and height. The View component contains a Text component with the numberOfLines and ellipsizeMode properties set to limit the number of lines of text that are displayed and to truncate the text if it extends beyond the screen. This will prevent the text error message from extending beyond the screen.